+++ CHANGELOG.md
@@ -8,6 +8,9 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 - check-ec2-cpu_balance.rb: scans for any t2 instances that are below a certain threshold of cpu credits
 - check-instance-health.rb: adding ec2 instance health and event data
 
+### Changed
+- Update to aws-sdk 2.2.11 and aws-sdk-v1 1.66.0
+
 ## [2.1.0] - 2016-01-15
 ### Added
 - check-elb-health-sdk.rb: add option for warning instead of critical when unhealthy instances are found